refactor(amp): moved amp dir inside server since amp is always SSR
