+++ server/amp/helpers/set-cors-headers.js
